Tidy up integral types in various re2/testing/*.cc files.

+++ re2/testing/re2_arg_test.cc
@@ -95,7 +95,7 @@ const int kNumStrings = arraysize(kSuccessTable);
   for (int i = 0; i < kNumStrings; ++i) {                                \
     RE2::Arg arg(&r);                                                    \
     const char* const p = kSuccessTable[i].value_string;                 \
-    bool retval = arg.Parse(p, strlen(p));                               \
+    bool retval = arg.Parse(p, static_cast<int>(strlen(p)));             \
     bool success = kSuccessTable[i].success[column];                     \
     EXPECT_EQ(retval, success)                                           \
         << "Parsing '" << p << "' for type " #type " should return "     \
